requirements.txt
aiogram==3.4.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiosqlite==0.20.0
python-dotenv==1.0.1
